Refactored un-used code for Illumina reads to misc package and made it IonTorrent Specific
